refactor: Lines로 클래스 분리
